rename branch dari updateStok ke updateHarga karena typo
