+++ src/View/UpdateStock.java
@@ -1,0 +1,19 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package View;
+
+/**
+ *
+ * @author glenn
+ */
+public class UpdateStock {
+    public static void main(String[] args) {
+        new UpdateStock();
+    }
+    public UpdateStock() {
+        
+    }
+}
